requirements.txt
numpy==1.26.4
scipy==1.13.0
auto-sklearn==0.15.0
xgboost
